fix(cloud): remove superseded workspace retire callback
